requirements.txt
python-telegram-bot==20.3
pandas==1.5.3
numpy==1.24.3
ta==0.10.2
pybit==5.9.0
httpx==0.24.1
